pyproject.toml
[project]
name = "pytest-loco-docs"
version = "1.3.3"
description = "Another one YAML-based DSL for testing"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]
authors = [
    {name = "Mikhalev Oleg", email = "mhalairt@gmail.com"}
]
license = "BSD-2-Clause"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "autodoc-pydantic (>=2.2.0,<3.0.0)",
    "furo (>=2025.12.19,<2026.0.0)",
    "pytest-loco (>=1.3.3,<2.0.0)",
    "pytest-loco-http (>=1.3.3,<2.0.0)",
    "pytest-loco-json (>=1.3.3,<2.0.0)",
    "sphinx (>=9.1.0,<10.0.0)",
    "sphinx-copybutton (>=0.5.2,<0.6.0)",
    "sphinx-inline-tabs (>=2025.12.21.14,<2026.0.0.0)",
]

[tool.poetry]
packages = [{include = "schemas", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
